[Java.Interop] Rename and re-namespace generic marshal methods
Fixes: https://github.com/xamarin/Java.Interop/issues/16

Commit f7750132 notes:

> This is nonsense, and further proof that the "generic marshaling" idea
> needs to be rethought, if only because you don't want to
> "accidentally" use them because of the overheads; see 35491137.

Commit 35491137 noted that the performance overhead of generic
marshaling was an added ~20-30% overhead while marshaling EACH
PARAMETER.

In short, generic marshaling should NOT be default, and even providing
it at all is dubious (though it does help for proof-of-concept and
robustness reasons, and generic marshaling should be altered so that
Java.Interop.Dynamic can use it).

Given the dubious nature of generic marshaling, the LEAST we can do is
prevent it from being used implicitly, accidentally, by default.

To do this, we do two things:

 1. Change the generic marshal methods to be extension methods,
    located in the Java.Interop.GenericMarshal namespace, and

 2. Rename the generic marshal methods so that "Generic" is
    in the name.

Thus, the previous code:

	using Java.Interop;
	partial class Demo {
		static readonly JniPeerMembers _members = new JniPeerMembers ("name", typeof (Demo));

		public int Example (int value)
		{
			_members.InstanceMethods.CallInt32Method ("example\u0000(I)I", this, value);
		}
	}

becomes:

	using Java.Interop;
	using Java.Interop.GenericMarshal;
	partial class Demo {
		static readonly JniPeerMembers _members = new JniPeerMembers ("name", typeof (Demo));

		public int Example (int value)
		{
			_members.InstanceMethods.CallGenericInt32Method ("example\u0000(I)I", this, value);
		}
	}

Note the addition of `using Java.Interop.GenericMarshal` and the use
of CallGenericInt32Method() instead of CallInt32Method().

This allows the reader to more quickly know which set of invocation
methods are being used, and reduces the liklihood of accidentally
using the generic marshaling mechanism when the explicit mechanism is
being used (allegedly; see also commit f7750132).