[build] Add XAIntegrationDebug configuration, `make xa-all` target.
As mentioned in commit a3469e88, we want to build a *subset* of
Java.Interop.dll for integration with Xamarin.Android, because I don't
feel that the entire Java.Interop API is stable, and thus shouldn't be
made public yet.

To that end, add an XAIntegrationDebug target to Java.Interop.csproj,
along with some MSBuild voodoo to selectively include only some of the
<Compile/> entries, not all of them, to (hopefully?) narrow down the
number of files that would otherwise require #if's to build.

(Meaning, I think it'll be easier to selectively *add* code instead of
selectively remove it.)

Add a `make xa-all` toplevel target to build just the
XAIntegrationDebug configuration outputs, which will be located in
$(topdir)/bin/XAIntegrationDebug.