[Java.Interop] Rename JavaVM to JniRuntime.
From commit 26a93217:

> I'm finding `JavaVM` more annoying the more I look at it

It's the "VM" part, specifically; it's irksome.

The "Java" prefix also violates convention (see README.md).

Rename JavaVM to JniRuntime. It's not *exactly* a runtime, but it's
not a VM either!

Additionally, rename JavaVMOptions to JniRuntime.CreationOptions, a
new (nested!) type. The rationale for making it nested is that almost
nobody needs this type, so I don't see any point in having it show up
in the Code Completion dropdown for `Java.Interop.`. Making it a
nested type hids it from the most common code-completion perspective.