[Java.Interop] Use more nested types for JniPeerMembers.
We're trying to fulfill multiple purposes here: have a "small"
Java.Interop.dll assembly, at least from the "code completion"
perspective, yet have a useful architectural split and functionality.

Which brings us to e.g. the JniPeerInstanceMethods type: it can't be
created alone, and can't be used in isolation. Yet it still shows up
in code completion, even though it should really never be used alone;
instead, it's intended to be used as part of JniPeerMembers:

	var members = new JniPeerMembers (...);
	members.InstanceMethods.InvokeVirtualVoidMethod (...);

To reduce the "code completion surface area" for toplevel *types*,
move the JniPeer* types as nested types of JniPeerMembers. This
removes them from type-level code completion, and reduces the number
of "public non-nested types" from 26 to 22 [0].

[0]: monop -r:bin/XAIntegrationDebug/Java.Interop.dll | grep '^Java.Interop\.' | grep -v '+' | wc -l