[Java.Interop] Expunge JniEnvironmentInfo from the public API.
Context: https://github.com/xamarin/Java.Interop/commit/26a93217fb41bac36754863b923102958d790182

JniEnvironmentInfo was only made public to make the JniEnvironment
refactoring (commit 26a93217) easier. Nothing in there need be part of
a public API.

> *something* per-thread needs to be updated for JNI local references...

That "something" need not be a JniEnvironmentInfo instance when it can
instead be a `ref int localReferenceCount`.

Using JniEnvironmentInfo would be useful if there were *additional*
information that we wanted to log alongside the JNI local reference
accounting, but at present I can't think of anything that would be
particularly relevant.

With this change, we are now down to 20 public types for the
XAIntegrationDebug configuration [0, 1].

[0]: Commit 11210c8a
[1]: monop -r:bin/XAIntegrationDebug/Java.Interop.dll | grep '^Java.Interop\.' | grep -v '+' | wc -l