[generator, Java.Interop] One TLS access per JNIEnv invocation.
Consider the non-P/Invoke based JNIEnv::NewIntArray():

	public static unsafe JniObjectReference NewIntArray (int length)
	{
		var tmp = JniEnvironment.Invoker.NewIntArray (JniEnvironment.EnvironmentPointer, length);
		JniEnvironment.LogCreateLocalRef (tmp);
		return new JniObjectReference (tmp, JniObjectReferenceType.Local);
	}

This requires *two* separate TLS accesses:

  * One for JniEnvironment.Invoker, and
  * One for JniEnvironment.EnvironmentPointer

TLS Access isn't necessarily "fast", or at least it's not necessarly
as fast a normal static variable access. Additionally, the two lookups
are *pointless*, because *both* of the values its looking up are
stored in the same JniEnvironmentInfo instance!

Capitalize on this, and use a single TLS lookup for both:

	public static unsafe JniObjectReference NewIntArray (int length)
	{
		var __info = JniEnvironment.CurrentInfo;
		var tmp = __info.Invoker.NewIntArray (__info.EnvironmentPointer, length);
		JniEnvironment.LogCreateLocalRef (tmp);
		return new JniObjectReference (tmp, JniObjectReferenceType.Local);
	}

The P/Invoke backend doesn't need this optimization, as its only
grabbing the JniEnvironment.EnvironmentPointer value.