[Java.Interop] Reduce jmethodID lookups.
A "funny" thing happened while working on integrating Xamarin.Android
and Java.Interop: I saw that the Java.Interop "efficient, cache
everything" virtual dispatch mechanism differed from what
Xamarin.Android was using for virtual-vs.-base dispatch:

	// Xamarin.Android:
	/* XA01 */ public override unsafe string ToString ()
	/* XA02 */ {
	/* XA03 */   if (id_toString == IntPtr.Zero)
	/* XA04 */     id_toString = JNIEnv.GetMethodID (class_ref, "toString", "()Ljava/lang/String;");
	/* XA05 */   if (GetType () == ThresholdType)
	/* XA06 */     return JNIEnv.GetString (JNIEnv.CallObjectMethod  (Handle, id_toString), JniHandleOwnership.TransferLocalRef);
	/* XA07 */   else
	/* XA08 */     return JNIEnv.GetString (JNIEnv.CallNonvirtualObjectMethod  (Handle, ThresholdClass, JNIEnv.GetMethodID (ThresholdClass, "toString", "()Ljava/lang/String;")), JniHandleOwnership.TransferLocalRef);
	/* XA09 */ }

	// Java.Interop JniPeerMembers.JniInstanceMethods.InvokeVirtual*Method():
	/* JI01 */ public unsafe JniObjectReference InvokeVirtualObjectMethod (string encodedMember, IJavaPeerable self, JniArgumentValue* parameters)
	/* JI02 */ {
	/* JI03 */   JniPeerMembers.AssertSelf (self);
	/* JI04 */   var declaringType   = DeclaringType;
	/* JI05 */   if (self.GetType () == declaringType || declaringType == null) {
	/* JI06 */     var m   = GetMethodInfo (encodedMember);
	/* JI07 */     return m.InvokeVirtualObjectMethod (self.PeerReference, parameters);
	/* JI08 */   }
	/* JI09 */   var j = self.JniPeerMembers;
	/* JI10 */   var n = j.InstanceMethods.GetMethodInfo (encodedMember);
	/* JI11 */   return n.InvokeNonvirtualObjectMethod (self.PeerReference, j.JniPeerType.PeerReference, parameters);
	/* JI12 */ }

In particular, Xamarin.Android XA05 checks ThresholdType, which is
equivalent to JniPeerMembers.ManagedPeerType, and is overridden in
every class binding. If they match, -- meaning the current type is a
binding type, *not* some custom user subclass -- then virtual dispatch
is used. Otherwise, we assume this is the user doing
`base.ToString()`, and we perform nonvirtual dispatch.

Java.Interop is slightly different; it compares self.GetType() --
equivalent to `GetType()` on XA05 -- to DeclaringType on JI05, but
that's it. Note, though, that DeclaringType is the *immediate* type of
the binding, and doesn't change for a given JniPeerMembers instance.

So let's provide an example: JavaObject overrides object.ToString() to
invoke java.lang.Object.toString().

Imagine that we have a java.lang.String binding, which doesn't
override ToString(), because it's not actually necessary:

	[JniTypeSignature (JniTypeName)]
	class MyString : JavaObject {
		internal    const   string      JniTypeName = "java/lang/String";

		internal    static  readonly    JniPeerMembers  _members    = new JniPeerMembers (JniTypeName, typeof (MyString));

		public override JniPeerMembers JniPeerMembers {
			get {return _members;}
		}

		public unsafe MyString (string value);
	}

Further imagine that we invoke `new MyString("value").ToString()`.

What happens is line JI05 fails, so Java.Interop hits the fallback
nonvirtual method invocation path, lines JI09-JI11.

While this works, and still produces the expected values, this
behavior isn't desirable because it means the jmethodID for
toString() needs to be looked up multiple times: one for JavaObject,
one for MyString, and one for *every other JavaObject subclass*, which
is *silly* because in *all* of these cases we're *not* doing
`base.ToSTring()`, and thus don't actually need a nonvirtual invoke.

The fix, then, is to relax the check on JI05 so that if
self.JniPeerMembers.ManagedPeerType matches self.GetType(), we use
virtual dispatch. This means we "share" the toString() jmethodID
lookup, reducing the number of jmethodIDs that need to be cached.

	- if (self.GetType () == declaringType || declaringType == null)
	+ if (self.GetType () == declaringType || declaringType == null || self.GetType () == self.JniPeerMembers.ManagedPeerType)