[Java.Interop] Move JniRuntime marshal-related methods to JniValueMarshaler
Related: https://github.com/xamarin/Java.Interop/commit/522ce7184e5a24d2d677ffd598245a562c8111f4
Related: https://github.com/xamarin/Java.Interop/commit/11210c8a

In the spirt of commit 522ce718, split out the instance bookkeeping
and marshaling code out from JniRuntime and move it into
JniRuntime.JniValueMarshaler.

The rationale is twofold:

 1. Reduce the conceptual surface area of JniRuntime, allowing fewer
    methods to be exposed yet allowing us to more sanely add methods
    in an orderly fashion without blowing up the core surface area.

 2. Allow "mix and match" implementations for JniRuntime subclasses.

An added possible benefit -- not currently used -- is that by
splitting it out into a separate type, we can separately test it
without needing to pull along an entire JniRuntime.

An instance of JniRuntime.JniValueMarshaler is available from the
JniRuntime.ValueMarshaler property, permitting use such as:

	var o = runtime.ValueMarshaler.GetObject<TargetType>(ref reference, options);

TODO:

  * Cleanup the JniValueMarshaler methods so there's a sane set of
    virtual methods for overriding, assuming such a thing is sane.

  * Determine the correct interraction between e.g. JavaObject and
    JniValueMarshaler and JniRuntime.

The latter is a big, unexplored, problem domain. JNI *permits* -- but
no *implmenetation* currently *supports* -- multiple JVMs loaded into
a single process. Meanwhile, there can be only one JNIEnv* pointer for
a given thread, from which you can access the corresponding JVM.

With that in mind -- specifically in the context of unit tests -- does
it "make sense" in future API revisions to:

  var m = new JniRuntime.JniValueMarshaler ();
  var o = new JavaObject ();
  m.Add (o);
  var n = m.GetObject (o.PeerReference);
  Assert.AreSame (o, n);

Conceptually, that makes a great deal of sense. Implementation-wise,
I'm not sure how that works, because JavaObject will be using the
JniRuntime.JniValueMarshaler associated with the current thread --
which *won't* be `m`! -- so logically we'd want to "move" the
JavaObject instance between the global
JniRuntime.Current.ValueMarshaler and `m`.

Except JavaObject doesn't (currently) have any reference to the
JniRuntime it was created with, much less the JniValueMarshaler, so
this quickly stops making any sense at all.

Even if JavaObject *did* have a reference to the "owning" runtime, in
the above "self-contained" scenario it wouldn't (can't) be registered
with the owning Runtime, so it *still* doesn't make sense to "move"
ownership in that fashion...does it?

Random food for thought...