[Java.Interop] Use ExceptionDispatchInfo to raise exceptions
(Small conceptual change; large patch. The "joy" of PCLs!)

The C# `throw` statement clears the Exception.StackTrace property,
re-filling it with the current stack trace at the time of the `throw`.

This is usually all well and good, until we have managed code calling
Java code calling managed code, and the managed code throws:

	Call stack: Managed ( > Java > Managed )+

Through JavaProxyThrowable, whatever exception instance was thrown by
the rightmost managed frame would be marshaled back to (each) managed
frame which calls a Java method, allowing e.g. an ArgumentException to
"percolate" back through all Java and managed callers, until handled
or the thread aborts.

Yay.

The problem is the code at the Managed > Java transition:

	Exception __e = JniEnvironment.GetExceptionForLastThrowable ();
	if (__e != null)
		throw __e;

Note that we're using `throw`, which means the *originally* preserved
stack trace will be *lost*, replaced with the current stack trace.

This makes for painful debugging experiences and sad pandas.
("Where was that exception thrown from?!")

.NET has a solution for this:
[System.Runtime.ExceptionServices.ExceptionDispatchInfo][0] has
Capture() and Throw() methods which will *update* the
Exception.StackTrace property without nuking the previous contents:

	Exception __e = JniEnvironment.GetExceptionForLastThrowable ();
	if (__e != null)
		ExceptionDispatchInfo.Capture (__e).Throw ();

This prevents the Exception instance from losing extremely useful and
important information, and makes for happier developers.

That's *all* we wanted to do.

Unfortunately, Java.Interop is PCL-based, and Profile6 -- the
currently used profile, based on .NET 4.0 -- doesn't contain
ExceptionDispatchInfo. We thus need to migrate to a .NET 4.5-based PCL
profile in order to improve matters. (Or delegate through JniRuntime,
but who wants to deal with that?)

Migrating to a .NET v4.5-based profile in turn requires migrating to
PCLv2 reference assemblies, which in turn *moves* lots of members that
were previously used:

  * Delegate.Method? Replaced with the Delegate.GetMethodInfo()
    extension method.

  * Delegate.CreateDelegate()? Replaced with
    MethodInfo.CreateDelegate().

  * Type.GetMethod()/GetConstructor()/etc.? Replaced with TypeInfo,
    which is available via the new Type.GetTypeInfo()
    extension method.

  * System.Threading.Thread? Fuck you, doesn't exist. *This* required
    "punting" by adding a new JniRuntime.GetCurrentThreadDescription()
    method, so that a "normal" .NET runtime can be used, e.g. the
    JreRuntime.GetCurrentThreadDescription() override.

This in turn required re-thinking e.g.
JniRuntime.JniObjectReferenceManager initialization, because
LoggingJniObjectReferenceManagerDecorator was unable to set
JniObjectReferenceManager.Runtime, resulting in
NullReferenceExceptions. Consequently, the internal
JniRuntime.ISetRuntime.SetRuntime() method is now
ISetRuntime.OnSetRuntime(), and is implemented as a *public* method
for all implementations, allowing
LoggingJniObjectReferenceManagerDecorator to delegate the call.

With all that done, we can now target the following PCL profiles:

  * Java.Interop.dll: Profile7 and/or Profile111. No others are easily
    usable because System.Runtime.Concurrent.dll and
    System.Runtime.InteropServices.dll are required.

  * Java.Interop.Dynamic.dll and Java.Interop.Export.dll: Profile78.
    Because...why not? They build. They won't *run* without
    Java.Interop.dll, but they build!

[0]: https://msdn.microsoft.com/en-us/library/system.runtime.exceptionservices.exceptiondispatchinfo(v=vs.110).aspx