[Java.Interop] API Review: Remove IntPtrs from JniEnvironment
Replace most IntPtr parameters within JniEnvironment with *real*
pointer types, e.g. for [JNIEnv::GetIntArrayElements()][0]:

	jint* GetIntArrayElements (JNIEnv *env, jintArray array, jboolean *isCopy);

The previous binding was:

	IntPtr GetIntArrayElements (JniObjectReference array, IntPtr isCopy);

Quick! Without consulting the documentation, what's the "real" return
type, so that you can actually *use* it? How is the `isCopy` parameter
supposed to be used?

System.IntPtr may be "safe" and "CLS Compliant", but it's
*meaningless*. The meaningful types are lost.

The `isCopy` parameter, for example, is "really" intended to be either
null, if you don't care, or the address of a `jboolean` OUT parameter.
`IntPtr` allows using this, but if you DO care, use is fugly:

	bool isCopy;
	JniEnvironment.Arrays.GetIntArrayElements (array, (IntPtr) (void*) &isCopy);

(Yes, the intermediate cast to `void*` is necessary; without it, you get
CS0118, which makes no sense at all.)

Then there's the return type: IntPtr means nothing, so what should it
be cast to for meaningful use? Again, hit the documentation -- which
still requires translating to managed types (though many of the
mappings are direct, e.g. `jint` to `int`).

Bleh.

Replace these IntPtr uses with unsafe, *real*, pointers. Thus:

	int* GetIntArrayElements (JniObjectReference array, bool* isCopy);

Want to pass `null` for `isCopy`? Still simple. Want to *get* the
output value for `isCopy`? Now it's saner:

	bool isCopy;
	int* elements = JniEnvironment.Arrays.GetIntArrayElements (array, &isCopy);

The return type? Trivial, and the compiler can better check for
consistency.

What's not to like?

[0]: http://docs.oracle.com/javase/8/docs/technotes/guides/jni/spec/functions.html#Get_PrimitiveType_ArrayElements_routines