[Java.Interop] Rethink JniObjectReferenceOptions
Consider JniObjectReferenceOptions:

	[Flags]
	enum JniObjectReferenceOptions {
		None                      = 0,
		CreateNewReference        = 1,
		DisposeSourceReference    = 2,
		DoNotRegisterWithRuntime  = 4,
	}

While [pondering enum naming][0], we came to the terrible conclusion
that JniObjectReferenceOptions semantics were poorly factored:

  @jonpryor:    CreateNewReference and DisposeSourceReference is
                either/or, not "and"
  @dellis1972:  yes you cant have CreateNewReference and
                DisposeSourceReference at the same time..
                it wouldn’t make sense

Which is to say, semantically, *all* of CreateNewReference,
DisposeSourceReference, and DoNotRegisterWithRuntime create a new
object reference; it's what is done *after* creating the reference
which changes. Because of that,
`CreateNewReference | DisposeSourceReference` doesn't really make
sense, yet it's a value which is permitted.

Let's try to improve this by "refactoring"/rethinking.

First of all, why do we want a [Flags] enum? For the
DoNotRegisterWithRuntime behavior, which *is* an "OR-able" flag.

Logically then, what we want is unique "operations" on the handle:

	[Flags]
	enum JniObjectReferenceOptions {
		None                      = 0,
		CreateNewReference        = 1,
		DisposeSourceReference    = 2,
		DoNotRegisterWithRuntime  = 4,
	}

Note the same names. ;-)

The problem with this is that the calling code gets annoying: if you
want to dispose the source reference, you need to specify everything:

	var s = JniEnvironment.Strings.ToString (
			ref r,
			JniObjectReferenceOptions.CreateNewReference | JniObjectReferenceOptions.DisposeSourceReference);

That's...*really* verbose.

Instead, becase ~100% of the time "DisposeSourceReference" and
"DoNotRegisterWithRuntime" imply/require CreateNewReference, let's
encode that into their values. We'll thus have this "implicit"/"full"
enumeration definition:

	[Flags]
	enum JniObjectReferenceOptions {
		None                      = 0,
		CreateNewReference        = 1,
		Copy                      = CreateNewReference,
		DisposeSourceReference    = 2,
		CopyAndDispose            = DisposeSourceReference | Copy,
		DoNotRegisterWithRuntime  = 4,
		CopyAndDoNotRegister      = DoNotRegisterWithRuntime | Copy,
	}

Then, for sanity, we remove the ones that shouldn't be used
separately, resulting in:

	[Flags]
	enum JniObjectReferenceOptions {
		None                      = 0,
		Copy                      = 1,
		CopyAndDispose            = 2 | Copy,
		CopyAndDoNotRegister      = 4 | Copy,
	}

This provides a small code-completion set with properly factored
semantics and no crazy-long invocation options:

	var s = JniEnvironment.Strings.ToString (ref r, JniObjectReferenceOptions.CopyAndDispose);

[0]: https://xamarinhq.slack.com/archives/android/p1447347043000479