[Java.Interop] Move JNI_GetCreatedJavaVMs() to JniRuntime
Remember a011cd7a, in which it was stated that JNI_GetCreatedJavaVMs()
and JNI_CreateJavaVM() didn't exist on Android, so those methods
should not be located in Java.Interop.dll?

Turns Out™ that advice is old: Android 6.0 with ART -- and maybe
earlier versions? -- *do* have these exports. JNI_CreateJavaVM()
errors out (unsurprisingly), but JNI_GetCreatedJavaVMs() works.

Thus, a question: can these be moved into JniRuntime?

JNI_GetCreatedJavaVMs() can be, so do so.

JNI_CreateJavaVM() is another matter, because to use
JNI_CreateJavaVM() you need JavaVMInitArgs and JavaVMOption, which in
turn imply things like `java.class.path` and various other runtime
settings, for which Android doesn't provide any (currently known)
facility.

In this case -- for now at least -- keep JNI_CreateJavaVM() within
Java.Runtime.Environment.dll, so that JreRuntimeOptions members don't
appear on Android-using environments.

---

Commit 967a7a8f asked:

>   * Determine the correct interraction between e.g. JavaObject and
>     JniValueMarshaler and JniRuntime.
>
> The latter is a big, unexplored, problem domain. JNI *permits* -- but
> no *implmenetation* currently *supports* -- multiple JVMs loaded into
> a single process. Meanwhile, there can be only one JNIEnv* pointer for
> a given thread, from which you can access the corresponding JVM.

There's a *bigger* problem here: multiple JVMs within a process
*isn't feasible*, not with the current architecture, and possibly not
ever. As such, we shouldn't support it, hence the new
Debug.WriteLine() WARNING message in this commit: You really,
*really*, shouldn't create more than one JniRuntime instance per
process.

Why?

Because of the GC.

A core architectural point -- not quite implemented here, but present
in Xamarin.Android -- is that JVM instances may reference Java.Interop
instances, and vice versa, and the GC should sort it all out.

Which it does! (Modulo GC bugs...)

The problem, as it pertains to "multiple JVMs in one process", is that
as per the JNI spec, a thread may be attached to only a single VM.

Which means the GC thread -- you know, the thing that needs to run to
collect instances and/or keep them alive! -- is *itself* attached to
ONLY ONE JVM.

Which means even if you *could* have more than one JVM in a process --
and no current JVM supports that! (commit 5009e28c) -- there'd be NO
WAY for the GC to *process* those objects, as it would treat all
instances as if they were from the same JVM, because it has to!

(This could plausibly be worked around by having JavaObject contain a
reference to the JniRuntime that "owns" it, and then the GC thread
could detach from one JVM and reattach to the other, but this is
(1) hopelessly complicated and subject to bugs, and (2) will have
terrible performance, assuming it would even work at all!)

(Further note that even if the above GC issue could be solved, we'd
also have to fix JniPeerMembers, because all of the
jmethodID/jfieldID/etc. metadata it caches is only valid for a
particular JVM. The values can't be used in any other JVM, meaning
we'd have to "move" much of the caching it currently performs into
e.g JniRuntime.JniTypeManager, which is more complexity & overhead.)