[Java.Interop.Dynamic] Fix JNI Method Signature Computation
One of the features in commit a9a4896b:

> we provide implementations for all builtin types, Nullable<T> for
> builtin types (which always box)

The primary intention for this was in the context of
Java.Interop.Dynamic, to "auto box" parameters:

	dynamic Integer = new DynamicJavaClass ("java/lang/Integer");
	dynamic value   = Integer(42);
	int     compare = value.compareTo(42);
	Assert.AreEqual (42, compare);

Reasonable enough for intent, but implementation-wise this gets
complicated because without parsing the JNI signature, there's no way
of knowing that we want an `int` passed or a `jobject` passed, and
JniValueMarshaler.CreateArgumentState()'s behavior varies.

There were two ways to address this:

 1. Parse the JNI signature to determine if we wanted an unboxed or
    boxed representation.

 2. Use a separate type, e.g. have nullable types be the boxed form.

The nice thing about (1) is that it allows the above fragment to work.
The cost is additional logic for parameter checking, logic which is
potentially wrong (yay bugs?).

Worse, to me, are the long-term hopes for tools/jnimarshalmethod-gen,
in which we use JniValueMarshaler to generate build-time marshaling
methods, and the JniValueMarshaler used would be based on the IL
parameter types, which in turn means we can't look at the runtime
parameter types in Java, as *there won't be a runtime JVM* then.

Which brings us to (2), which is what was done in a9a4896b: treat
nullable types of the builtin types as object references.

This allows nearly identical code to work:

	dynamic Integer = new DynamicJavaClass ("java/lang/Integer");
	dynamic value   = Integer (42);
	int     compare = value.compareTo((int?) 42);
	Assert.AreEqual (42, compare);

Forcing the use of `(int?)` causes us to use the
JniValueMarshaler<int?> implementation, which boxes into
java.lang.Integer instances, allowing us to invoke
java.lang.Integer.compareTo(java.lang.Integer), as desired.
Even better, you can tell at the call site what is intended.

Why wasn't this test added with commit a9a4896b? Because it didn't
work, and I didn't want to track down why it didn't work at the time.

I have since done so.  The problem is that
JavaMethodBase.LookupArguments() was treating arguments improperly: it
was simply appending the result of
JniEnvironment.Types.GetJniTypeNameFromClass().  The problem with this
is that for java.lang.Integer it'll return `java/lang/Integer`, while
within a JNI signature we instead need `Ljava/lang/Integer;`.  Using
the wrong representation meant that JNIEnv::GetMethodID() failed.

Fix JavaMethodBase.LookupArguments() to instead use
JniTypeSignature.QualifiedReference, which Does The Right Thing™,
which allows our intended JniValueMarshaler + `dynamic` integration
test to actually work.